requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
aiohttp==3.9.0
    # via
    #   discord-py
    #   langchain
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
anyio==3.7.1
    # via langchain
async-timeout==4.0.3
    # via
    #   aiohttp
    #   langchain
attrs==23.1.0
    # via aiohttp
certifi==2023.11.17
    # via requests
charset-normalizer==3.3.2
    # via requests
dataclasses-json==0.6.2
    # via langchain
discord==2.3.2
    # via -r requirements.in
discord-py==2.3.2
    # via discord
exceptiongroup==1.2.0
    # via anyio
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.0.1
    # via sqlalchemy
idna==3.4
    # via
    #   anyio
    #   requests
    #   yarl
jsonpatch==1.33
    # via langchain
jsonpointer==2.4
    # via jsonpatch
langchain==0.0.340
    # via -r requirements.in
langsmith==0.0.66
    # via langchain
marshmallow==3.20.1
    # via dataclasses-json
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
numpy==1.26.2
    # via langchain
packaging==23.2
    # via marshmallow
pydantic==2.5.2
    # via
    #   langchain
    #   langsmith
pydantic-core==2.14.5
    # via pydantic
python-dotenv==1.0.0
    # via -r requirements.in
pyyaml==6.0.1
    # via langchain
requests==2.31.0
    # via
    #   langchain
    #   langsmith
sniffio==1.3.0
    # via anyio
sqlalchemy==2.0.23
    # via langchain
tenacity==8.2.3
    # via langchain
typing-extensions==4.8.0
    # via
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
urllib3==2.1.0
    # via requests
yarl==1.9.3
    # via aiohttp
